aiXplain Certification Project/requirements.txt
requests==2.31.0
loguru==0.7.2
chromadb==0.4.10
sentence-transformers==2.2.2
streamlit==1.28.1
python-dotenv==1.0.0
numpy==1.24.3
scikit-learn==1.3.1
pandas==2.1.1
